Day1: better late than never
Got started a couple days late due to work + setting up a new laptop.
Going to at least try out some lua this year. The first part was simple
and a nice intro kind of project.
The second was unexpectedly difficult. While doing a bit of refactoring
I'd accidentally stopped storing the numeric value in the table and
instead had the string. Due to the way they were printed out it was
actually pretty unclear it was happening, which seems like a bit of a
knock against lua.