day15: try using a data structure
Tables are swell and all, but after finishing this with an
implementation that took 10 minutes to run I decided that perhaps
iterating over hundreds of cells to find the min every time was actually
a little bit too inefficient. I found a heap implementation online,
converted it to fennel, and then slightly rewrote it so that I could
reinsert the same key without doing the up or down pushing
implementations (which might have actually been easier to add). The
runtime is now ~0.5s, which is significantly nicer.

Also added the bold green to more easily see the path, and the bold red
because green wasn't distinct enough in my terminal...